LayoutTests/http/tests/security/mixedContent/insecure-basic-auth-image-UpgradeMixedContent.https.html

<!DOCTYPE html><!-- webkit-test-runner [ UpgradeMixedContentEnabled=true ] -->
<html>
<head>
<script src="/js-test-resources/js-test.js"></script>
<script>
if (window.testRunner) {
    testRunner.setHandlesAuthenticationChallenges(true);
    testRunner.setAuthenticationUsername("testUser");
    testRunner.setAuthenticationPassword("testPassword");
}

window.jsTestIsAsync = true;

function pass()
{
    testPassed("did not load image.");
    finishJSTest();
}

function fail()
{
    testFailed("did load image.");
    finishJSTest();
}

function runTest()
{
    // Load the image programmatically instead of declaratively to avoid output flakiness caused by
    // the preload scanner performing mixed content checks as part of preloading the image.
    let image = new Image;
    image.onload = fail;
    image.onerror = pass;
    image.src = "http://localhost:8000/security/mixedContent/resources/subresource2/protected-image.py";
    document.body.appendChild(image);
}

window.onload = runTest;
</script>
</head>
<body>
<script>
description("Tests that we do not load an insecure image when asked for basic authentication credentials.");
</script>
</body>
</html>
